# swagger/Docs.yaml
openapi: 3.0.0
info:
  title: Team 695 API
  description: Team 695 API Documentation for multiple services including authentication, events, surveys, teams, uploads, and feedback
  version: 1.5.2
servers:
  - url: https://api.team695.com
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Also called Access Key
    apiSecretAuth:
      type: apiKey
      in: header
      name: X-API-Secret
      description: Used together with X-API-Key, also called Access Secret Key
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        message:
          type: string
          description: Optional message providing additional information
        data:
          type: object
          description: Response data (varies by endpoint)
        error:
          type: string
          description: Error details (only in non-production environments)
...
    UploadResponse:
      type: object
      properties:
        url:
          type: string
          description: URL of the uploaded file
    EventResponse:
      type: object
      properties:
        eventId:
          type: string
          description: ID of the current event
...
    CreateAssignmentRequest:
      type: object
      properties:
        event_key:
          type: string
          description: Event key (e.g. 2025ohcl)
        task_type:
          type: string
          enum: [scouting, pit-scouting]
          description: Type of scouting task
        assigned_team_numbers:
          type: array
          items:
            type: integer
          description: List of team numbers for pit-scouting
        assigned_alliance:
          type: string
          enum: [red, blue]
          description: Alliance color for scouting
        assigned_matches:
          type: string
          description: Range of matches for scouting (e.g. Q3-Q12)
        assignees_data:
          type: array
          items:
            $ref: '#/components/schemas/UserData'
          description: List of users assigned to this task
        notes:
          type: string
          description: Additional notes for the assignment
      required:
        - event_key
        - task_type
        - assignees_data
    UpdateAssignmentRequest:
      type: object
      properties:
        assigned_team_numbers:
          type: array
          items:
            type: integer
          description: List of team numbers for pit-scouting
        assigned_alliance:
          type: string
          enum: [red, blue]
          description: Alliance color for scouting
        assigned_matches:
          type: string
          description: Range of matches for scouting (e.g. Q3-Q12)
        assignees_data:
          type: array
          items:
            $ref: '#/components/schemas/UserData'
          description: List of users assigned to this task
        status:
          type: string
          enum: [pending, in_progress, completed, canceled]
          description: Current status of the assignment
        notes:
          type: string
          description: Additional notes for the assignment
    TaskAssignment:
      type: object
      properties:
        id:
          type: integer
          description: Assignment ID
        event_key:
          type: string
          description: Event key (e.g. 2025ohcl)
        task_type:
          type: string
          enum: [scouting, pit-scouting]
          description: Type of scouting task
        assigned_team_numbers:
          type: array
          items:
            type: integer
          description: List of team numbers for pit-scouting
        assigned_alliance:
          type: string
          enum: [red, blue]
          description: Alliance color for scouting
        assigned_matches:
          type: string
          description: Range of matches for scouting (e.g. Q3-Q12)
        assigner_data:
          $ref: '#/components/schemas/UserData'
        assignees_data:
          type: array
          items:
            $ref: '#/components/schemas/UserData'
          description: List of users assigned to this task
        status:
          type: string
          enum: [pending, in_progress, completed, canceled]
          description: Current status of the assignment
        notes:
          type: string
          description: Additional notes for the assignment
        created_at:
          type: string
          format: date-time
          description: Assignment creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Assignment last update timestamp
...
    UserData:
      type: object
      properties:
        email:
          type: string
          description: User email address
        avatar:
          type: string
          description: URL to user avatar
        userId:
          type: string
          description: User ID
        username:
          type: string
          description: Username
        displayName:
          type: string
          description: User display name
...
paths:
...
  /upload/upload:
    post:
      summary: Upload a file
      description: Upload an image file to the server
      tags:
        - Uploads
      security:
        - bearerAuth: []
        - apiKeyAuth: []
          apiSecretAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: File to upload
                type:
                  type: string
                  description: Type of upload (e.g., 'robots', 'drivetrain')
      responses:
        '200':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UploadResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /event/event-id:
    get:
      summary: Get current event ID
      description: Retrieve the ID of the most recent event (public endpoint)
      tags:
        - Events
      responses:
        '200':
          description: Event ID retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EventResponse'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /assignments:
    post:
      summary: Create a new task assignment
      description: Create a new scouting or pit-scouting task assignment (admin only)
      tags:
        - Assignments
      security:
        - bearerAuth: []
        - apiKeyAuth: []
          apiSecretAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssignmentRequest'
      responses:
        '200':
          description: Assignment created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TaskAssignment'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Forbidden - Admin privileges required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /assignments/{id}:
    put:
      summary: Update a task assignment
      description: Update an existing task assignment (admin only)
      tags:
        - Assignments
      security:
        - bearerAuth: []
        - apiKeyAuth: []
          apiSecretAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Task assignment ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAssignmentRequest'
      responses:
        '200':
          description: Assignment updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TaskAssignment'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Forbidden - Admin privileges required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Assignment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      summary: Delete a task assignment
      description: Delete an existing task assignment (admin only)
      tags:
        - Assignments
      security:
        - bearerAuth: []
        - apiKeyAuth: []
          apiSecretAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Task assignment ID
      responses:
        '200':
          description: Assignment deleted successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          deleted:
                            type: boolean
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Forbidden - Admin privileges required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Assignment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...